+++ demo/src/main/java/com/example/demo/controller/CustomerController.java
@@ -23,7 +23,7 @@ public class CustomerController {
     }
 
     @GetMapping("/{id}")
-    public CustomerGetDTO getCustomer(@PathVariable Long id){
+    public CustomerGetDTO getCustomer(@PathVariable Integer id){
         return customerService.getById(id);
     }
 
@@ -33,12 +33,12 @@ public class CustomerController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Customer> updateCustomer(@PathVariable Long id, CustomerUpdateDTO customer){
+    public ResponseEntity<Customer> updateCustomer(@PathVariable Integer id, CustomerUpdateDTO customer){
         return customerService.update(id, customer);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Customer> deleteCustomer(@PathVariable Long id){
+    public ResponseEntity<Customer> deleteCustomer(@PathVariable Integer id){
         return customerService.delete(id);
     }
 }
