+++ demo/src/main/java/com/example/demo/dto/CityDTO.java
@@ -1,4 +1,4 @@
 package com.example.demo.dto;
 
-public record CityDTO(String city, Short country_id) {
+public record CityDTO(String city, Integer country_id) {
 }
